+++ arena/workbench/__init__.py
@@ -1,0 +1,1 @@
+"""Code workbench helpers: managed runtimes and project workspaces."""